backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
alembic>=1.13
aiosqlite>=0.20
playwright>=1.48
playwright-stealth>=2.0
cryptography>=43.0
httpx>=0.27
python-telegram-bot>=21.0
pydantic-settings>=2.5
python-multipart>=0.0.9
